+++ frontend/src/components/Create/Create.jsx
@@ -29,6 +29,7 @@ export default class Create extends Component {
         room: info.room,
         playScene: info.play,
         pauseScene: info.pause,
+        pauseDelayMs: info.pauseDelayMs ?? 0,
         stopScene: info.stop,
         resumeScene: info.resume,
         scrobble: info.scrobble,
@@ -68,6 +69,7 @@ export default class Create extends Component {
         playScene: "-1",
         playRoom: "-1",
         pauseScene: "-1",
+        pauseDelayMs: 0,
         stopScene: "-1",
         resumeScene: "-1",
         scrobble: "-1",
@@ -180,6 +182,7 @@ export default class Create extends Component {
     temp.play = this.state.playScene;
     temp.stop = this.state.stopScene;
     temp.pause = this.state.pauseScene;
+    temp.pauseDelayMs = this.state.pauseDelayMs;
     temp.resume = this.state.resumeScene;
     temp.scrobble = this.state.scrobble;
     temp.scrobbleDelayMs = this.state.scrobbleDelayMs;
@@ -305,6 +308,12 @@ export default class Create extends Component {
   handlePause = (e) => {
     this.setState({
       pauseScene: e.target.value.toString(),
+    });
+  };
+
+  handlePauseDelay = (e) => {
+    this.setState({
+      pauseDelayMs: e.target.value.toString(),
     });
   };
 
@@ -614,6 +623,41 @@ export default class Create extends Component {
               ))}
             </Form.Select>
             <div className="div-seperator" />
+            {/* Pause Delay */}
+            <Form.Label for="pauseDelay">
+              Pause Delay (ms) &nbsp;&nbsp;
+              <OverlayTrigger
+                placement="right"
+                overlay={
+                  <Tooltip>
+                    This will delay the pause action after receiving the webhook for the selected amount of time unless
+                    a resume action happens before the delay time has expired. This delay is intended for users using
+                    Plex clients that send out a pause webhook when using the seek feature.
+                    <br />
+                    <br />
+                    Be aware that this will also delay normal pause actions. It is advised to try and find a delay time
+                    that will satisfy the seek issue but not take too long on normal pauses.
+                  </Tooltip>
+                }
+              >
+                <img src={Info} className="image-info" alt="Info" />
+              </OverlayTrigger>
+            </Form.Label>
+            <Stack gap={1} direction="horizontal">
+              <Form.Range
+                id="pauseDelay"
+                className="me-auto"
+                value={this.state.pauseDelayMs}
+                min={0}
+                max={5000}
+                step={100}
+                onChange={this.handlePauseDelay}
+              />
+              <div className="slider-style" style={{ whiteSpace: "nowrap" }}>
+                {this.state.pauseDelayMs == 0 ? "Off" : this.state.pauseDelayMs + " ms"}
+              </div>
+            </Stack>
+            <div className="div-seperator" />
             {/* Resume Action */}
             <Form.Label for="resume">
               Resume &nbsp;&nbsp;
@@ -708,7 +752,7 @@ export default class Create extends Component {
                 onChange={this.handleScrobbleDelay}
               />
               <div className="slider-style" style={{ whiteSpace: "nowrap" }}>
-                {this.state.scrobbleDelayMs} ms
+                {this.state.scrobbleDelayMs == 0 ? "Off" : this.state.scrobbleDelayMs + " ms"}
               </div>
             </Stack>
             <div className="div-seperator" />